requirements.txt
praw
pyyaml
openai